% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/standardError.R
\name{se}
\alias{se}
\title{Calculates standard errors}
\usage{
se(x)
}
\arguments{
\item{x}{Vector of values for which the standard error is calculated}
}
\description{
This function calculates the standard error of a vector of values
}
\examples{
test_data <- c(2.3, 1.9, 3.2, NA, 1.7, 2.0, 2.6)
standard_error <- se(test_data)
}
\keyword{error}
\keyword{standard}

